refactor(uibackend): rename uibackendclient package to client
